Remove version from output + remove reading from disk
Havng the output include the version was going to cause more problems
that it's worth, so removed it.

Additionally, the reading from files was not going to be tennable for the
generator, so removed that suppor premptively. (We'll switch to assuming
the generator is good, and manually generating the output).