Support duplicate class disambiguation (e.g. suffix it)
When generating from a resw files, the _name_ of the file (and thus the
resource map) may not be a legal class name. To address that, the name
is cleaned up so that it's a legal class name. **However**, multiple
files could end up with the same name (`Waffle.en-us.resw` and
`Waff.le.en-us.resw` are both `Waffle`), which is not valid either.

So we track them for a given instance of the `WrapperGenerator` and
suffix the class name if it's the second or higher class of the same name